Move package to contributor
