+++ android/src/test/java/io/github/droidkaigi/feeder/ContributorViewModelTest.kt
@@ -1,7 +1,7 @@
 package io.github.droidkaigi.feeder
 
-import io.github.droidkaigi.feeder.core.ContributorViewModel
-import io.github.droidkaigi.feeder.core.fakeContributorViewModel
+import io.github.droidkaigi.feeder.contributor.ContributorViewModel
+import io.github.droidkaigi.feeder.contributor.fakeContributorViewModel
 import io.kotest.matchers.ints.shouldBeGreaterThan
 import io.kotest.matchers.types.shouldBeInstanceOf
 import kotlinx.coroutines.InternalCoroutinesApi
