+++ uicomponent-compose/other/src/main/java/io/github/droidkaigi/feeder/contributor/ContributorViewModel.kt
@@ -1,10 +1,11 @@
-package io.github.droidkaigi.feeder.core
+package io.github.droidkaigi.feeder.contributor
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.compositionLocalOf
 import io.github.droidkaigi.feeder.AppError
 import io.github.droidkaigi.feeder.Contributor
+import io.github.droidkaigi.feeder.core.UnidirectionalViewModel
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.StateFlow
